Fix problem with Context: Create component UserContext Accep component in index.tsx
