+++ .prettierrc
@@ -4,6 +4,5 @@
   "printWidth": 100,
   "singleQuote": true,
   "trailingComma": "none",
-  "jsxBracketSameLine": true,
-  "no-unused-vars":false
+  "jsxBracketSameLine": true
 }
